# backend-laravel/src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ApiPost - linuxitos",
        "description": "API con login, registro, actualización de información de usuario, perfil, listar posts, agregar, actualizar y eliminar posts. Obtiene token de inicio de sesión y se valida temporalidad.",
        "contact": {
            "email": "contact@linuxitos.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/v1/",
            "description": "Servidor local de desarrollo"
        }
    ],
    "paths": {
...
        "/api/v1/user/update": {
            "post": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Actualizar perfil del usuario autenticado",
                "operationId": "aac6294085079e6659c990fadc85422b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Juan Pérez"
                                    },
                                    "password_current": {
                                        "type": "string",
                                        "example": "claveactual123"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "nuevaclave123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "example": "nuevaclave123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Perfil actualizado correctamente"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/user/avatar": {
            "post": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Actualizar avatar del usuario",
                "operationId": "0cc36c2b9fccaf2f09c1188d05520e58",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "avatar"
                                ],
                                "properties": {
                                    "avatar": {
                                        "description": "Archivo de imagen (jpg, png, webp)",
                                        "type": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Avatar actualizado con éxito"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/user/mypermits": {
            "post": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Obtener permisos del usuario autenticado",
                "operationId": "07dfcf34579543cba9a54fa5dd5a3c47",
                "responses": {
                    "200": {
                        "description": "Permisos obtenidos"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/user/assignpermit": {
            "post": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Asignar o revocar permiso a un usuario",
                "operationId": "7902fec9173ac156d236c68effbd9623",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id",
                                    "sub"
                                ],
                                "properties": {
                                    "id": {
                                        "description": "ID del usuario",
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "sub": {
                                        "description": "ID del submódulo",
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "status": {
                                        "description": "true para asignar, false para revocar",
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Permiso actualizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/users": {
            "post": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Obtener lista de usuarios o uno específico",
                "operationId": "56b2198f4dbcffb2f3f482752edc633f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "order_by",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listado obtenido"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/users/delete": {
            "post": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Eliminar usuario por ID",
                "operationId": "49faf613aa1d6a4c2633fc45af9939dd",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": 3
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuario eliminado correctamente"
                    },
                    "404": {
                        "description": "Usuario no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}